--- a/agent/Cargo.toml
+++ b/agent/Cargo.toml
@@ -78,6 +78,7 @@ kube = { version = "0.74", default-features = false, features = [
 ] }
 nix = "0.23"
 openshift-openapi = { version = "0.3.1", features = ["v4_6"] }
+pcap = "0.9.1"
 
 [patch.crates-io]
 openshift-openapi = { git = "https://github.com/deepflowys/openshift-openapi.git" }
@@ -99,7 +100,6 @@ windows_recv_engine = { path = "plugins/windows_recv_engine" }
 [dev-dependencies]
 criterion = "0.3.5"
 lru = "0.7.5"
-pcap = "0.9.1"
 tempfile = "3.2.0"
 uluru = "3.0.0"