--- a/agent/crates/public/src/consts.rs
+++ b/agent/crates/public/src/consts.rs
@@ -186,7 +186,7 @@ pub const TCP_OPT_FLAG_SACK: u8 = 0x38; // 0011 1000, 同时也表示SACK的字�
 pub const TCP_OPT_WIN_SCALE_LEN: usize = 3;
 pub const TCP_OPT_MSS_LEN: usize = 4;
 
-// IPVS: https://github.com/yubo/ip_vs_ca 
+// IPVS: https://github.com/yubo/ip_vs_ca
 pub const TCP_OPT_ADDRESS_IPVS: u8 = 200;
 // HUAWEI & ALI: https://github.com/Huawei/TCP_option_address/blob/master/src/toa.h
 pub const TCP_OPT_ADDRESS_HUAWEI: u8 = 254;